Business central import settings PR comments fix
