+++ src/app/core/models/business-central/business-central-configuration/business-central-import-settings.model.ts
@@ -1,7 +1,6 @@
 import { FormArray, FormControl, FormGroup } from "@angular/forms";
-import { ImportSettingMappingRow } from "../../common/import-settings.model";
+import { ImportSettingMappingRow, ImportSettingsModel } from "../../common/import-settings.model";
 import { IntegrationField } from "../../db/mapping.model";
-import { RxwebValidators } from "@rxweb/reactive-form-validators";
 
 export type BusinessCentralImportSettings = {
     import_categories: boolean,
@@ -19,65 +18,8 @@ export interface BusinessCentralImportSettingsPost extends BusinessCentralImport
 
 export class BusinessCentralImportSettingsModel {
 
-    static createFormGroup(data: ImportSettingMappingRow): FormGroup {
-        return new FormGroup ({
-          source_field: new FormControl(data.source_field || '', RxwebValidators.unique()),
-          destination_field: new FormControl(data.destination_field || '', RxwebValidators.unique()),
-          import_to_fyle: new FormControl(data.import_to_fyle || false),
-          is_custom: new FormControl(data.is_custom || false),
-          source_placeholder: new FormControl(data.source_placeholder || null)
-        });
-    }
-
-    static constructFormArray(importSettings: null | BusinessCentralImportSettingsGet, businessCentralFields: IntegrationField[]): FormGroup[] {
-        const expenseFieldFormArray: FormGroup[] = [];
-        const mappedFieldMap = new Map<string, any>();
-        const unmappedFieldMap = new Map<string, any>();
-
-        // First loop to populate mappedFieldMap
-        businessCentralFields.forEach((businessCentralField) => {
-            const mappingSetting = importSettings?.mapping_settings.find(
-                (setting) => setting.destination_field === businessCentralField.attribute_type
-            );
-
-            const fieldData = mappingSetting || {
-                destination_field: businessCentralField.attribute_type,
-                import_to_fyle: false,
-                is_custom: false,
-                source_field: '',
-                source_placeholder: null
-            };
-            if (mappingSetting) {
-                mappedFieldMap.set(businessCentralField.attribute_type, fieldData);
-            } else {
-                unmappedFieldMap.set(businessCentralField.attribute_type, fieldData);
-            }
-
-        });
-
-        // Handle only mapped fields
-        businessCentralFields.forEach((businessCentralField) => {
-            const fieldData = mappedFieldMap.get(businessCentralField.attribute_type);
-            if (fieldData) {
-                expenseFieldFormArray.push(this.createFormGroup(fieldData));
-            }
-        });
-
-        if (mappedFieldMap.size === 0){
-            businessCentralFields.forEach((businessCentralField) => {
-                if (expenseFieldFormArray.length < 3) {
-                const fieldData = unmappedFieldMap.get(businessCentralField.attribute_type);
-                if (fieldData) {
-                    expenseFieldFormArray.push(this.createFormGroup(fieldData));
-                }
-                }
-            });
-        }
-        return expenseFieldFormArray;
-    }
-
     static mapAPIResponseToFormGroup(importSettings: BusinessCentralImportSettingsGet | null, businessCentralFields: IntegrationField[]): FormGroup {
-        const expenseFieldsArray = this.constructFormArray(importSettings, businessCentralFields);
+        const expenseFieldsArray = ImportSettingsModel.constructFormArray(importSettings, businessCentralFields);
         return new FormGroup({
             importCategories: new FormControl(importSettings?.import_categories ?? false),
             expenseFields: new FormArray(expenseFieldsArray)
